requirements.txt
paho_mqtt==1.6.1
pydantic==1.10.4
PyQt5==5.15.9
